Templatise BST (still only testing int)
WHAT!? Templated class functions need to be defined in the header
file??? This is crazy news to me, that I can no longer have separate
BST.h and BST.cpp files. If you had them in separate files, the compiler
gives you unresolved external symbol errors.
It's got something to do with the fact that these functions are not
instantiated until they are used in some other cpp file. That cpp file
will only #include the .h file, so if the function definition is not in
there, then it cannot instantiate the function. Big oof.